Clock Applet: port to modern GNOME technologies
This is a big commit that updates the Clock Applet to use modern
libraries. This includes using GSettings instead of GConf, using
GDateTime instead of the equivalents from libc and using the new
API in libgweather.

https://bugzilla.gnome.org/show_bug.cgi?id=677512

+++ applets/clock/clock-map.c
@@ -417,7 +417,7 @@ clock_map_mark (ClockMap *this, gfloat latitude, gfloat longitude, gint mark)
 static gboolean
 clock_map_place_location (ClockMap *this, ClockLocation *loc, gboolean hilight)
 {
-        gfloat latitude, longitude;
+        gdouble latitude, longitude;
 	gint marker;
 
         clock_location_get_coords (loc, &latitude, &longitude);
